avatarUrl

+++ src/controllers/author.controller.ts
@@ -1,6 +1,8 @@
 import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "express";
 const authorClient = new PrismaClient().author;
+
+
 // getAll
 export const getAllAuthors = async (req: Request, res: Response) => {
   try {
@@ -28,6 +30,10 @@ export const getOneAuthor = async (req: Request, res: Response) => {
         books: true,
       },
     });
+    if (!author) {
+      res.status(404).json({ error: "Author not found" });
+      return;
+    }
     res.status(200).json({ data: author });
   } catch (error) {
     console.log(error);
@@ -38,6 +44,10 @@ export const getOneAuthor = async (req: Request, res: Response) => {
 export const createAuthor = async (req: Request, res: Response) => {
   try {
     const authorData = req.body;
+    const file = req.file;
+    if(file){
+      authorData.avatarUrl = `http://localhost:3001/uploads/${file.filename}`;
+    }
     const newAuthor = await authorClient.create({
       data: authorData,
     });
@@ -51,14 +61,29 @@ export const createAuthor = async (req: Request, res: Response) => {
 export const updateAuthor = async (req: Request, res: Response) => {
   try {
     const authorData = req.body;
+    console.log(authorData)
+    const file = req.file;
+    if(file){
+      authorData.avatarUrl = `http://localhost:3001/uploads/${file.filename}`;
+    }
     const authorId = parseInt(req.params.id);
-    const author = await authorClient.update({
+    const author = await authorClient.findUnique({
+      where: {
+        id: authorId,
+      },
+    });
+    if (!author) {
+      res.status(404).json({ error: "Author not found" });
+      return;
+    }
+    const updatedAuthor = await authorClient.update({
       where: {
         id: authorId,
       },
       data: authorData
     });
-    res.status(200).json({ data: author });
+
+    res.status(200).json({ data: updatedAuthor });
   } catch (error) {
     console.log(error);
     res.status(400).json({ error: "Invalid request" });
@@ -69,11 +94,21 @@ export const updateAuthor = async (req: Request, res: Response) => {
 export const deleteAuthor = async (req: Request, res: Response) => {
   try {
     const authorId = parseInt(req.params.id);
-    const author = await authorClient.delete({
+    const author = await authorClient.findUnique({
       where: {
         id: authorId,
       },
     });
+    if (!author) {
+      res.status(404).json({ error: "Author not found" });
+      return;
+    }
+    await authorClient.delete({
+      where: {
+        id: authorId,
+      },
+    });
+
     res.status(204).json("Author has been successfully deleted");
   } catch (e) {
     console.log(e);
