+++ src/controllers/book.controller.ts
@@ -30,6 +30,10 @@ export const getOneBook = async (req: Request, res: Response) => {
         id: bookId,
       }
     });
+    if (!book) {
+      res.status(404).json({ error: "Book not found" });
+      return;
+    }
     res.status(200).json({ data: book });
   } catch (error) {
     console.log(error);
@@ -59,7 +63,16 @@ export const updateBook = async (req: Request, res: Response) => {
   try {
     const { title, description: description, datePublished, authorId }: Book = req.body; 
     const bookId = parseInt(req.params.id);
-    const book = await bookClient.update({
+    const book = await bookClient.findUnique({
+      where: {
+        id: bookId,
+      },
+    });
+    if (!book) {
+      res.status(404).json({ error: "Book not found" });
+      return;
+    }
+    const updatedBook = await bookClient.update({
       where: {
         id: bookId,
       },
@@ -70,7 +83,7 @@ export const updateBook = async (req: Request, res: Response) => {
         author: { connect: { id: authorId } }
       }
     });
-    res.status(200).json({ data: book });
+    res.status(200).json({ data: updatedBook });
   } catch (error) {
     console.log(error);
     res.status(400).json({ error: "Invalid request" });
@@ -81,7 +94,16 @@ export const updateBook = async (req: Request, res: Response) => {
 export const deleteBook = async (req: Request, res: Response) => {
   try {
     const bookId = parseInt(req.params.id);
-    const book = await bookClient.delete({
+    const book = await bookClient.findUnique({
+      where: {
+        id: bookId,
+      },
+    });
+    if (!book) {
+      res.status(404).json({ error: "Book not found" });
+      return;
+    }
+    await bookClient.delete({
       where: {
         id: bookId,
       },
